New top-level package: wrench
This package provides simple way for any part of Juju to check if a
"wrench should be dropped in the works" at some part of Juju's
execution. It checks $datadir/wrench for files which signal that
something should break, performing permission and error checking, and
logging.

The idea is that judicious calls to wrench.IsActive can be permanently
left in the code to aid manual and CI testing.