Removed invalid peergrouper test
An old test that checks that the peergrouper doesn't start with the
local provider was still present even though the peergrouper is now
always run. The reason this test has been passing is because the
peergrouper doesn't start until upgrades are finished (the test was
passing for the wrong reason). Now that upgrade mode isn't entered
unless it is necessary, this test started failing.