cmd/jujud: avoid potential race for UpgradedToVersion
Store the initial UpgradedToVersion as the the machine agent starts up
and use that when deciding whether HA initialisation is
required. Previously it could have been possible for the upgrade-steps
worker to update UpgradedToVersion before ensureMongoServer got to
check it.

The new previousAgentVersion field will also be useful for other
upcoming changes.