Make environs/bootstrap responsible for tools location
environs/bootstrap now finds available tools, augmenting
or substituting for locally buildable/uploadable tools
as necessary (auto upload) or as specified (--upload-tools).
Environ.Bootstrap is now provided with a list of available
tools, and is not responsible for syncing tools itself.

--upload-series and --series will be obsoleted: we will
upload a single series, and then use that in the metadata
for each supported series.