cmd/juju: move some tests to their own package to reduce the overall test time.
This horrible PR is a response to the cmd/juju tests now taking over 600s to complete. This causes the build to fail, sometimes the build will pass with `-p 2`, but more commonly this will not happen.

Given that all the tests use the odeous JujuConnSuite there is little that can be done to make them run faster, each test takes 3-4 seconds to run `SetUpTest` on my machine on a ram drive, I can only imagine how long they take to run IN THE CLOUD.

Yes, this PR is horrible, and will mean pretty soon we'll need a cmd/juju/test2, etc. This is the cost of not having proper mocks.