state/api: always "upload" all supported series
UploadTools no longer accepts a "fakeSeries"
parameter; the client just sends all of the
supported series of the same OS as the uploaded
tools. This is to support old API servers.

A separate commit will change the API server
side to disregard the parameter, and compute
the supported series on the server side. We
can deprecate the client side parameter in a
future version.

Also: return a generic error if status code
is not 200. We no longer need to check the
"method not supported" error specifically.