state/apiserver: move api/params to apiserver/params
This PR moves api/params to apiserver/params. The justification is the params package defines the form of the api messages that the server will accept, not the form that the client will provide.

This move makes it clear that the parameters belong to the apiserver, not a Go implementation of the api client.