state: remove "authenticated", simplify txnRunner/getManagedStorage
Don't need the "authenticated" logic around session copying anymore,
as we now run mongo without auth in the tests, and in production we
always open state with a tag/password.

txnRunner now returns the session associated with the runner, and
getManagedStorage and ToolsStorage are updated to use that.
txnRunnerWith is dropped.