Search, download and cache tools from public/configured tools sources
FindTools will now search simplestreams unless an exact match is
being performed and that is found in toolstorage. Tools found in
toolstorage will take precedence over those in simplestreams.

If a tools version is requested for download and it is not in
toolstorage already, the API server will search for it in simplestreams,
download it, and cache it in toolstorage before returning it to
the client.

Also, we now once again (as in 1.20) only clone tools on upload if
requested. We will drop the "AddToolsAlias" function from toolstorage.