api: api no longer depends on state.
As part of a previous refactoring I accedently made api directly dependent on state, which is an even worse crime than state depending on the apiserver.

This proposal corrects this mistake and removes all dependencies on state from the api packages, along with additional removals of direct dependencies on state.

The cost is the agent package is responsible for translating between params.StateServingInfo, which is effectively the public data type of state information, and state.StateServingInfo which is the state's own version of the same.