cmd/juju: upgrade-juju "upgrade in progress" handling
upgrade-juju now special-cases handling the case of the
SetEnvironAgentVersion API call failing due to an upgrade appearing to
already be in progress. The user is given guidance on what they might
be able to do to fix it.

The change introduces a struct conforming to the upgradeJujuAPI
interface to allow testing of the upgrade-juju command without making
actual API calls.