state: Life no longer depends on apiserver/params
`state.Life` is a small integer, however it relies on params.State for its string representation. Apart from the coupling, this is nuts because common/life.go doesn't check the results of `state.Life.String()`, it just converts it to a `params.Life` and stuffs it back to the client.