upgrades: read charms from disk in local/manual
Local and manual provider storage is httpstorage
layered on filestorage serving from a location
specified in agent config. The storage is served
by a "localstorage" worker, which does not start
until after upgrades complete. We shortcut the
httpstorage and go straight to disk if the provider
is local or manual.