apiserver: allow an extra API method during upgrades
Client.EnvironmentGet is required to support "juju ssh". SSH should be
available during upgrades but this method was not in the list of
upgradingRoot's allowed API methods.

This revision fixes the issue (tested manually). A CI test change will
be made next ensure that SSH continues to work during upgrades.

Fixes LP#1367009.