state/multiwatcher: break dependency on params.EntityInfo
This change breaks a dependency from `state/multiwatcher` to `apiserver/params` for the `EntityInfo` type. This is done by declaring a compatible interface in the `multiwatcher` package. As two interfaces with identical method sets are themselves identical, values which fulfil the `params.EntityInfo` interface also fulfil `multiwatcher.EntityInfo`.

Sadly `EntityInfo` defines a `EntityId` method which returns a `params.EntityId` type, which continues to wed us to `apiserver/params`. That problem is the subject of subsequent PRs