decouple params.EntityInfo.EntityId from multiwatcher.EntityInfo.EntityId
This change uses a dirty hack to _temporarily_ break the dependency between the two types that communicate using an EntityInfo interface type.

You don't have to like it, just understand that it isn't permenant, just a step to avoid a 4,000 line change that inverts the dependencies between these two packages.