worker/uniter/jujuc: Fix TestAddMetric on gccgo
Change logic of add-metric command to avoid relying on map ordering
in tests, which is undefined and different on gccgo. Rather than
recording metrics as a map, keep it as a slice, but use a map when
parsing to check and error still on duplicate keys.

Also shortens error message on duplicate keys and makes output
from TestAddMetrics clearer.