Possible fix for Bug #1304126
In the SetEnvironmentCommand's Init function I added an additional API call to retrieve existing environment config.
All keys a user specifies in the "juju set-env" command are checked against the existing config keys and if the
user defined key does not exist in the existing config a warning of a possible misspelling is issued.