A proposed fix for bug #1384622
Added a check to the "juju set" command, which issues a warning to the user
if the value of the configuration option is already set to the desired value.

The check is done by retrieving the unit's configuration options via an
additional API call and comparing the user-defined value with the
existing configuration value.