api/uniter: fixed ordering assumption in TestJoinedRelations
This test used to sort the received tags before comparing to the
expected results but a recent revision changed this. The test was only
passing by luck because the returned tags are ordered according to
MongoDB's default undefined ordering. The env UUID update for
relations changed the returned ordering making this test fail.

jc.SameContents is now used so the returned ordering is irrelevant.