Force the requests to actually be concurrent by adding a Barrier.
The code wanted to make sure that we would have concurrent requests to Put, so we make sure to
pause Put() until all of the requests have been started.
Adding a time.Sleep(10ms) to the goroutine loop causes it to fail reliably (because the
goroutines end up effectively serialized), but if we add the putBarrier it makes them
concurrent again even with the sleeps.