Fixes bug #1364939
This pull request, together with https://github.com/juju/juju/pull/1026, fixes the bug #1364939.

- Added a destryo call in lxc.go if a container fails to start. If the destroy call succeeds the
CreateContainer returns a RetryableCreationError, which means that the provisioner may safely
retry container creations. If the destroy call fails, this means that there is a leftover container
in an error state, which could not be destroyed and may affect subsequent container creations - an
intervention from the user is required to clean up the environment (using lxc-destroy).
- Added test using the mock-lxc - and error injection channel was added which induces errors in the Start
and Destroy functions.
- Fixed the CreateContainerTest in container/testing/common.go (errors.Trace masks the true error type)