Fixes bug #1364939
Depends on changes in https://github.com/juju/juju/pull/1026

- Added a cleanup step in lxc.go if a containers fails to start. If the lxcContainer.Destroy() call does
not return an error a RetryableCreationError is returned, which means that it is safe for the provisioner
to retry instance creation (changes in the provisioner are in PR #1026). If the Destroy call fails
a RetryableCreationError should not be returned - a container is left in an error state and user
intervention may be required to clean up the environment.
- Added an error injection channel to the mock lxc code - if a message is received on the error injection
channel the Start or Destroy function will fail. This is used to test the changes in lxc.go.
- Added tests for the code introduced to lxc.go in lxc_test.go (see TestCreateContainerFailsWithInjectedError
and TestCreateContainerWithInjectedErrorDestroyFails)
- Fixed the CreateContainerTest in container/testing/common.go (errors.Tracef masks the true type of
the error)