Do not dump the oplog.
The oplog is a component of mongo replication.  It represents idempotent
events in the database system. [1]  The --oplog flag to mongodump causes
any oplog entries added after the dump starts to be written out to
oplog.bson at the end of the dump.  It provides a point-in-time snapshot
guarantee about the dump.  This matters under replication since other
nodes in the replica set could propagate operations while the dump is
running.  However, if mongo is running without replication then there is
no oplog to dump.

For now we are going to dump the databases without the oplog.
Subsequent patches will address dumping the oplog and addressing ignored
databases relative to it.

[1] See:
  http://www.kchodorow.com/blog/2010/10/12/replication-internals/
  http://docs.mongodb.org/manual/tutorial/troubleshoot-replica-sets/
  http://www.tokutek.com/2014/06/why-tokumx-changed-mongodbs-oplog-format-for-operations/