Remove the new backups CLI.
While the rest of the backups implementation is complete, a key
component (restore) has not landed in time for 1.21.  The existing
restore plugin *should* work with the new backups code, but that has
not been tested at all.  Regardless, it doesn't make a lot of sense
to mix the old and new backups in that way.

Consequently we are effectively disabling the new backups capability
in 1.21.  Completely excising the new backups implementation would
introduce too much risk.  Removing just the CLI portion is sufficient
to effectively eliminate exposing the new backups implementation to
users.  The new API facade is still in place and the various packages
are still there in the code, but that is not likely to be an issue.

This code will remain in the master branch (for 1.22 and forward).