provider/ec2: allocate all instance stores
When starting an instance in EC2, specify all of the
instance stores in the block device mapping. The instance
started will have as many four attached, according to
the instance type.

getBlockDeviceMappings will be enhanced in the near
future to support attaching additional volumes to
instances at creation time (i.e. for charm storage).