state: fixed service life handling in SetMetricCredentials
This method was adding the env UUID prefix onto an id that already had
the prefix when checking service life. This meant the life check was
never going to succeed and the associated test was passing for the
wrong reason (service not found instead of service is dying/dead).

The function was also only checking for the Dead state (but not
Dying). Based on the test, I have assumed that SetMetricCredentials is
only supposed to work if a service is alive so I've changed the
implementation to check for Alive instead of Dead.