Move state/storage.go to state/storage/
This moves storage (blobstore) code out of state,
and into the state/storage package. The storage code
is inconsequential to state itself, it just happens
to be co-resident in the Mongo database.

The moved code does not depend on state, but instead
takes an environment UUID and an mgo.Session derived
from a state.State.

In the future we should also move the ToolStorage
code, removing the dependency of state on blobstore
altogether.