Remove storage filesystem preferences
The spec has been updated, removing the charm author's
ability to specify preferred filesystems. For advanced
use-cases, charms will be required to request a block
device and format it.

The diskformatter has been simplified to just create
an ext4 filesystem. The choice of filesystem can be
revised later if necessary. We still record filesystem
type (currently always ext4) and mount options (currently
always empty) so we can make adjustments with minimal
disruption.

The storage.Specification struct has been removed for
now; it will be reintroduced when it is needed again,
with up-to-date contents mirroring the spec.