Store params for requested block devices in state
We introduce a feature-flag protected "--disks" flag to
"juju machine add", which will pass a set of storage
constraints to the AddMachine call. These constraints
will be converted to parameters for disk provisioning
and recorded in state, with a reference to the machine.

A follow-up branch will update the machine provisioner
so that it retrieves these requested block devices and
passes the parameters to StartInstance. The instance
broker will create the disks and return information
about them to then store in state along with the other
instance info.