Introduce charm storage instances
We introduce a feature-flag protected flag to
"juju deploy" called "--storage", which enables
users to specify storage constraints for charm-
defined stores. The constraints are passed to
the API server which will validate them against
the charm metadata and the available storage
providers, and then record the constraints on
the service in state.

When a unit is added to a service with storage
constraints, the unit will instantiate storage
instances (except for shared stores, which are
not yet supported; for these there will be one
instance for the entire service). Unit
destruction cannot now succeed if there are
live storage instances associated with the unit,
to prevent leaking resources.

When a unit is assigned to a machine, block
devices will be requested for each of the unit's
storage instance, and the requested block
devices will be assigned to those storage
instances. When a storage instance is removed
from state, the block devices will be unassigned
from the storage instance, but destruction is
optional.