all: revise storage model
The storage model is revised to cater for shared
storage, including shared filesystems and volumes.

We now model "attachments", which store machine
and unit-specific information relating to a volume
or storage instance. For example, volumes may be
multiply attached (depending on provider support),
so the block device name is stored on the volume
attachment.

Machine agents will watch for volume and filesystem
entities and attachments, and will (de)provision
them as necessary. Units will watch for storage
attachments, which will be updated when the
corresponding volume or filesystem is.