apiserver: close State once connection closed.
Using resources to clean up a State connection used by the API means
that the State could be closed while it's still being used, causing
panics. The new Cleaner interface is now used to close State instead
(when required).

Cleanup was implemented on both apiRoot and apiHandler because either
one may be used for serving a connection.