cmd/jujud: fix panic due to prematurely closed state
When returning with an error, startEnvWorkers wasn't waiting for the
runners it created to be done before returning. This meant that
sometimes the envWorkerManager would close the State that was still in
use by workers and pingers attached to those runners, causing panics.

This change also removes an unnecessary goroutine that was being used
to kill the singular runner created by startEnvWorkers. Singular
runners stop themselves when the underlying runner stops so there was
no need for this.