State: Make presence watcher and pinger mulit-env aware
Pass in the environment UUID to presence.NewWatcher and NewPinger.
Use the UUID to ensure .beings .seqs and .pings collections are
environment specific. Test that the watcher sees ping events
isolated to its environment. Move the setup of the presence
watcher for state to a place where we have the environment UUID
available.