Remove juju/names package
The juju/names package was used to hold four constants to assist with managing the need for a .exe suffix for windows binaries.

The package is too heavy handed a solution for a the problem. As part of its it crowded the package import space; multiple files were already importing a package named 'names', and in all but one case only one of the constants defined in the package was used.

A lighter weight solution is to deal with the two specific cross platform issues at the point they are encountered

* if a command is being executed, ensure it has ".exe" appended when compiled on windows.
* if a tool is being referenced by name, for copying, for example, again make sure it grows a ".exe" suffix