cmd/jujud/agent: use valid config when adding extra environments
TestNewEnvironmentStartsNewWorkers was using an environ config that
the dummy provider didn't consider valid. The firewaller worker waits
for a valid environment and was getting stuck when the machine agent
was being stopped. Whether the problem was visible or not is highly
timing dependent but was consistently occuring on some CI hosts.

Fixes LP #1416577.