Fix restore/azure
Azure was returning ErrNotBootstrapped from its StateServerInstances
method, which is not necessarily true. In the restore scenario run
by QA, Azure is bootstrapped and then the state server instance is
killed OOB. So, we change the StateServerInstances method to return
ErrNotBootstrapped instead. The "EnsureNotBootstrapped" code still
assumes that the environment is not bootstrapped if this error is
returned from StateServerInstances (i.e. the conversion moves up a
level).

The Azure provider will now also handle the scenario where there
are cloud services without deployments, returning ErrNoInstances
instead of ErrPartialInstances.

We also update the restore plugin to check for: ErrPartialInstances
(multiple state server instances created, some but not all alive),
and ErrNoInstances in the result of StateServerInstances.

Fixes https://bugs.launchpad.net/juju-core/+bug/1417178