Safeguard destroy state server environment and remove all env docs
Only allow the state server environment to be destroyed after all other
environments are destroyed.

Add a new method on State, RemoveAllEnvironDocs, which removes all
documents from all collections in state.multiEnvCollections as well as
the environments document associated with the given state. The method
is intended to be called after a non-state-server environment is
destroyed.

Call RemoveAllEnvironDocs at the end of
"apiserver/client".Client.DestroyEnvironment(), when we are sure all
cleanup steps have been taken and only if the destroyed environment was
not the state-server.