Implement loop storage provider
The loop storage provider allocates files in the agent's
data-dir ($data-dir/storage/loop), and attaches loop
devices to them. This gives us a universal storage
provider for block devices.

There is a change to the storage.Provider.VolumeSource
method signature: it now takes a storage directory.
This is a directory that will be provided by the storage
provisioner, based on data-dir, for the storage provider
to store local artifacts. It should only be used by
machine-scoped storage providers.

Also, drive-by simplification of VolumeAttachment,
removing the VolumeId and InstanceId fields. We only
need to convey those in the params, we don't ever need
to inform Juju of the volume ID or instance ID when
creating an attachment.