Two changes to envworkermanager.
1) We don't need runner.Dying() we can use the paradigm used elsewhere.
Instead of selecting on the Dying() channel, just start a goroutine that
curries the runner's Wait() into our tomb. (so if the runner exits, our
tomb will get the message and cause us to exit.) This will let us remove
Dying from being a member of the Runner interface.

2) If there is any reason that *this* processing exits, then it should Kill()
the runner. The easiest way to trigger this today is to make GetEnvironment fail
with an error (potentially an Unauthorized, though NotFound is currently handled).