Move get / set constraints into common directory
This patch does three things:

1 - Moves the get/set-constraints code into the
common directory (see below).  The addition of the
get/set-constraints subcommands for environment
and service will be added as a separate commit.

2 - Removes the test dependency on mongoDB

3 - Changes get/set-constraints to be registered
as deprecated in 2.0 and obsolete in 3.0

As part of the hierarchical command work, there are
a few commands which can be under more than one super
command.  For example, get-constraints can be accessed
by both:

juju environment get-constraints
juju service get-constraints

To help organize these commands, the core functionality
will be moved into the juju/juju/cmd/juju/common
directory.  When possible, test dependencies on mongoDB
will be removed.