Merge pull request #1616 from dimitern/lp-1416928-filter-lxcbr0-addresses
Fixed lp:1416928 - ignore lxcbr0 addresses on machine-0 for non-local environments

The address(es) of the LXC bridge device is detected and filtered before
setting API hosts:ports (on bootstrap and when changes are detected by
the apiaddressesupdater) and also when updating machine addresses
discovered from the machiner locally. This means that bootstrapping on
a machine with the lxc package installed (i.e. having lxcbr0) or
installing the lxc package later won't cause the state or API addresses
to change and include the lxcbr0 address(es), which are unusable in all
but local environments.

See http://pad.lv/1416928 for more info.