Move ensure-availability back to a top level command
It has been decided that ensure-availability will
eventually move under a new juju supercommand: server

Until that new supercommand is put in place, the
ensure-availability command should remain at the
top level to avoid confusion.

I kept the changes made to mock out the API for
testing, and moved the end-to-end test for
ensure-availability into the featuretests directory.