state: add Filesystem bits.
And some related cleanups of volumes, defaults.

Adding a filesystem consults the storage provider
to determine whether or not the filesystem requires
a Juju-managed volume backing it, and adds one if
required. The Filesystem interface provides a
method for obtaining the tag of the backing volume
if there is one.