New juju environment get/set-constraints subcommand
The core functionality for get/set-constraints resides
in the juju/juju/cmd/juju/common directory, and implements
constraints commands for both environments and services.

The get/set-constraints subcommands for environment only
need to enforce that service constraints are not set or
retrieved through that command.

The docs for these subcommands reference the soon-to-be-
created service supercommand, which will be in a separate
patch later this week.