New service supercommand with get/set-constraints subcommands
This patch introduces a the new service supercommand under
juju.  The new service get/set-constraints subcommands are
also included.

As with the environment get/set-constraints subcommands,
the core functionality for get/set-constraints resides
in the juju/juju/cmd/juju/common directory, and implements
constraints commands for both environments and services.

The get/set-constraints subcommands for service only
need to enforce that only service constraints are set or
retrieved through those commands.

The behavior of service set-constraints was also modified
from the behavior of the top level set-constraints command
to not require a -s or --service flag for specifying the
service.  For consistency, both get and set expect that the
first argument passed in is the service name.