CMD: environment unshare
Add "juju environment unshare" sub-command and corresponding tests in
cmd, api and featuretests. When unshare is called on a user that has
not been added to an environment, the APIServer errors but the CLI
only logs a warning - as the desired outcome has still been reached:
the user does not have access to the environment.

(Review request: http://reviews.vapour.ws/r/1037/)