State: Ensure user cannot have two environments with same name
Add a new collection, userenvnameC, to use in a txn op which places a
unique key restraint on the "owner" and "name" fields of
environmentsC. Update state.NewEnvironment(): if an environment cannot
be found after insertion check if it is because of the unique key.

Ensure the userenvnameC doc is added and removed with the environment.