state: use errors.Cause in IsNotAssigned
There was a change in PR 1707 that causes the
megawatcher to fail if the unit is not yet assigned to
a machine. The cause is that a call to errors.Cause was
removed in back-porting a fix from master, but
IsNotAssigned does not do this in 1.22 as it does on
master.

This branch changes IsNotAssigned to call errors.Cause.

Fixes https://bugs.launchpad.net/juju-core/+bug/1430049