State: Make username checks case insensitive
Add a new "name" field to  state.userDoc and migrate the value of
"_id" to it. Then lower case "_id". Ensure no more that one user can be
created with the same, case insensitive, name and that state.User()
searches with a lowercased name. Update userenvnameC to
store a lower cased username.