api: extract connection logic out of Open
The code that establishes a websocket connection to one of the API
servers as been extracted out of Open into a separate Connect
func. Most of the tests for Open were actually exercising this part of
Open so they have been adjusted to match.

This will soon be used by the database logging worker.